backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-marshmallow
marshmallow-sqlalchemy
flask-cors
python-dotenv
psycopg2-binary
bcrypt
requests
google-generativeai
